Refactoring: MockRetrievedNode
